fix: run cancelled event remediation status check only for mongodb
