+++ tests/fault_management_test.go
@@ -19,8 +19,6 @@ package tests
 
 import (
 	"context"
-	"fmt"
-	"strings"
 	"testing"
 	"time"
 
@@ -30,8 +28,6 @@ import (
 	"github.com/stretchr/testify/require"
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
-	"k8s.io/client-go/rest"
-	"sigs.k8s.io/e2e-framework/klient"
 	"sigs.k8s.io/e2e-framework/pkg/envconf"
 	"sigs.k8s.io/e2e-framework/pkg/features"
 
@@ -563,7 +559,7 @@ func TestManualUncordonWithFaultRemediation(t *testing.T) {
 
 		client, err := c.NewClient()
 		require.NoError(t, err)
-		mongoPod = helpers.GetMongoDBPrimaryPodName(ctx, t, client)
+		mongoPod, _ = helpers.TryGetMongoDBPrimaryPodName(ctx, t, client)
 
 		return newCtx
 	})
@@ -650,12 +646,17 @@ func TestManualUncordonWithFaultRemediation(t *testing.T) {
 	})
 
 	feature.Assess("verify FR marks cancelled event remediated", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
+		if mongoPod == "" {
+			t.Log("Skipping direct faultRemediated MongoDB assertion for non-MongoDB datastore")
+			return ctx
+		}
+
 		client, err := c.NewClient()
 		require.NoError(t, err)
 
 		t.Log("Waiting for FR to write faultRemediated=true on the Cancelled HealthEvent")
 		require.Eventually(t, func() bool {
-			return faultRemediatedMarkerWritten(ctx, t, client.RESTConfig(), client, mongoPod,
+			return helpers.CancelledHealthEventFaultRemediated(ctx, t, client.RESTConfig(), client, mongoPod,
 				testCtx.NodeName, cancellationMarkerMessage)
 		}, helpers.EventuallyWaitTimeout, helpers.WaitInterval,
 			"FR should write faultRemediated=true for the cancelled HealthEvent")
@@ -679,37 +680,4 @@ func TestManualUncordonWithFaultRemediation(t *testing.T) {
 	})
 
 	testEnv.Test(t, feature.Feature())
-}
-
-func faultRemediatedMarkerWritten(
-	ctx context.Context,
-	t *testing.T,
-	restConfig *rest.Config,
-	client klient.Client,
-	mongoPod, nodeName, message string,
-) bool {
-	t.Helper()
-
-	js := fmt.Sprintf(`
-		db = db.getSiblingDB("%s");
-		const doc = db.HealthEvents.findOne({
-			"healthevent.nodename": %q,
-			"healthevent.message": %q,
-			"healtheventstatus.nodequarantined": "Cancelled"
-		});
-		if (doc && doc.healtheventstatus &&
-			doc.healtheventstatus.faultremediated &&
-			doc.healtheventstatus.faultremediated.value === true) {
-			print("FAULT_REMEDIATED_TRUE");
-		} else {
-			print("NOT_READY");
-			if (doc) {
-				printjson(doc.healtheventstatus);
-			}
-		}
-	`, helpers.MongoDBDatabase, nodeName, message)
-
-	stdout, _ := helpers.ExecMongosh(ctx, t, restConfig, client, mongoPod, js)
-
-	return strings.Contains(stdout, "FAULT_REMEDIATED_TRUE")
 }
